ml/mydjangoapp/requirements.txt
Django
scikit-learn
joblib
pandas
numpy
mysqlclient
djangorestframework